Crate flux_common

Modules

bug

cache

dbg

This file contains functions and macros to log debugging information meant for developers.

index

iter

mir_storage

This module allows storing mir bodies with borrowck facts in a thread-local storage. Unfortunately, thread local storage requires all data stored in it to have a 'static lifetime. Therefore, we transmute the lifetime 'tcx away when storing the data. To ensure that nothing gets meessed up, we require the client to provide a witness: an instance of type TyCtxt<'tcx> that is used to show that the lifetime that the client provided is indeed 'tcx.

result

Macros

_basic_block_start

_check_fn_span

_debug_assert_eq3

_hyperlink

_refine_goto

_refine_mode_span

_shape_goto_enter

_shape_goto_exit

_shape_mode_span

_statement

_terminator

bug

span_bug

tracked_span_assert_eq

tracked_span_bug

tracked_span_dbg_assert_eq